+++ packages/svelte-undp-design/src/lib/Select/Select.stories.ts
@@ -14,6 +14,9 @@ const meta = {
 		},
 		items: {
 			description: 'The list of items of SelectItem interface'
+		},
+		selectedItem: {
+			description: 'Selected item'
 		}
 	}
 } satisfies Meta<Select>;
@@ -21,18 +24,29 @@ const meta = {
 export default meta;
 type Story = StoryObj<typeof meta>;
 
+const items = [
+	{ label: 'All', value: 'default' },
+	{ label: 'Pasto', value: 'pasto' },
+	{ label: 'Dari', value: 'dari' },
+	{ label: 'English', value: 'english' },
+	{ label: 'Albanian', value: 'albanian' },
+	{ label: 'Arabic', value: 'arabic' },
+	{ label: 'Portuguese', value: 'portuguese' }
+];
+
 // More on writing stories with args: https://storybook.js.org/docs/7.0/svelte/writing-stories/args
 export const Primary: Story = {
 	args: {
 		placeholder: 'Category',
-		items: [
-			{ label: 'All', value: 'default' },
-			{ label: 'Pasto', value: 'pasto' },
-			{ label: 'Dari', value: 'dari' },
-			{ label: 'English', value: 'english' },
-			{ label: 'Albanian', value: 'albanian' },
-			{ label: 'Arabic', value: 'arabic' },
-			{ label: 'Portuguese', value: 'portuguese' }
-		]
+		items: items,
+		selectedItem: undefined
+	}
+};
+
+export const Selected: Story = {
+	args: {
+		placeholder: 'Category',
+		items: items,
+		selectedItem: items[items.length - 1]
 	}
 };