+++ sites/geohub/src/components/maplibre/symbol/IconOverlap.svelte
@@ -1,7 +1,6 @@
 <script lang="ts">
 	import { MAPSTORE_CONTEXT_KEY, type MapStore } from '$stores';
 	import { getContext } from 'svelte';
-	import RangeSlider from 'svelte-range-slider-pips';
 
 	const map: MapStore = getContext(MAPSTORE_CONTEXT_KEY);
 
@@ -9,36 +8,21 @@
 
 	const choices = ['never', 'always', 'cooperative'];
 	const propertyName = 'icon-overlap';
-	let selected = [
-		choices.findIndex((choice) => choice === $map.getLayoutProperty(layerId, propertyName))
-	];
+	let selected = $map.getLayoutProperty(layerId, propertyName) ?? choices[0];
 
 	$: selected, setIconOverlap();
 
 	const setIconOverlap = () => {
-		map.setLayoutProperty(layerId, propertyName, choices[selected[0]]);
+		map.setLayoutProperty(layerId, propertyName, selected);
 	};
 </script>
 
-<div data-testid="icon-overlap-slider" class="range-slider" style="width: 130px;">
-	<RangeSlider
-		bind:values={selected}
-		min={0}
-		formatter={(v) => choices[v].toLowerCase().replace(/\b(\w)/g, (s) => s.toUpperCase())}
-		max={choices.length - 1}
-		all="label"
-		pips
-	/>
+<div data-testid="icon-overlap-slider" class="select is-fullwidth">
+	<select bind:value={selected}>
+		{#each choices as choice}
+			<option value={choice}
+				>{choice.toLowerCase().replace(/\b(\w)/g, (s) => s.toUpperCase())}</option
+			>
+		{/each}
+	</select>
 </div>
-
-<style lang="scss">
-	.range-slider {
-		--range-handle-focus: #2196f3;
-		--range-range-inactive: #2196f3;
-		--range-handle-inactive: #2196f3;
-		--range-handle: #2196f3;
-		width: 100%;
-		cursor: pointer;
-		font-size: 11px;
-	}
-</style>