+++ sites/geohub/src/routes/(app)/data/upload/+page.svelte
@@ -68,7 +68,7 @@
 		const shapefileFiles = zipFilesList.map((file) => {
 			// check if the file has a valid shapefile extension
 			const extension = file.path.split('.').at(-1);
-			if (shapefileExtensions.includes(extension)) {
+			if (shapefileExtensions.includes(extension.toLowerCase())) {
 				return file.path;
 			}
 		});
@@ -138,7 +138,7 @@
 			const extension = file.name.split('.').at(-1);
 			if (
 				AccepedExtensions.find((ext) => ext.name === 'ESRI Shapefile').extensions.includes(
-					extension
+					extension.toLowerCase()
 				)
 			) {
 				// return file if not undefined
@@ -330,6 +330,7 @@
 		input.accept = AccepedExtensions.map((ext) =>
 			ext.extensions.map((e) => `.${e}`).join(',')
 		).join(',');
+
 		input.click();
 		input.onchange = async (e) => {
 			let files = e.target.files;
@@ -399,7 +400,9 @@
 				];
 				return;
 			}
-			const formats = AccepedExtensions.filter((ext) => ext.extensions.includes(extension));
+			const formats = AccepedExtensions.filter((ext) =>
+				ext.extensions.includes(extension.toLowerCase())
+			);
 			if (formats.length === 0) {
 				errorMessages = [...errorMessages, `The file extension '${extension}' is not supported.`];
 				return;
@@ -472,14 +475,16 @@
 			</label>
 		</div>
 	</Dropzone>
-	<div class="is-normal is-flex is-align-items-center mt-5 is-justify-content-space-between">
-		<span>
-			Click
-			<DefaultLink title="here" href="/data/supported-formats" target="_blank" />
-			to read about supported formats
-		</span>
-		<div class="is-normal is-flex is-align-items-center">
-			<div class="ml-2 help">
+	<div class="columns mt-5 is-justify-content-space-between">
+		<div class="column is-flex-mobile">
+			<span>
+				Click
+				<DefaultLink title="here" href="/data/supported-formats" target="_blank" />
+				to read about supported formats
+			</span>
+		</div>
+		<div class="column">
+			<div class="is-flex is-align-items-center help">
 				<Checkbox
 					disabled={!userIsSignedIn || isUploading}
 					on:clicked={() =>
@@ -487,15 +492,16 @@
 					checked={!config.DataPageIngestingJoinVectorTiles}
 					label="Every layer (Point, Line, Polygon) into its own file"
 				/>
+				<Help>
+					Most of GIS data formats can hold more than one vector layer. The option below, if checked
+					will result in extracting each layer as a different dataset (own metadata, name, and other
+					properties). The alternative is to join all layers into one multi-layer dataset where
+					layers are hidden inside and not discoverable directly.
+				</Help>
 			</div>
-			<Help>
-				Most of GIS data formats can hold more than one vector layer. The option below, if checked
-				will result in extracting each layer a different dataset (own metadata, name and other
-				properties). The alternative is to join all layers into one multi-layer dataset where layers
-				are hidden inside and not discoverable directly.
-			</Help>
 		</div>
 	</div>
+
 	{#if filesToUpload.length > 0}
 		<div class="table-container mt-5">
 			{#if filesToUpload.length > 0}
@@ -519,7 +525,11 @@
 							<tr>
 								<td>
 									<div>
-										<span>{path ? path.split('.').at(-2) : name.split('.').at(-2)}</span>
+										<span
+											>{path
+												? path.split('.').slice(0, -1).join('.')
+												: name.split('.').slice(0, -1).join('.')}</span
+										>
 										{#if path}
 											<span class="tag is-medium is-info is-light"
 												>.{path ? path.split('.').at(-1) : name.split('.').at(-1)}</span
@@ -593,11 +603,15 @@
 											<span class="tag is-grey-light">{uploadStatusMapping[name]}</span>
 										{/if}
 									</td>
-									<td>
-										{#if !uploadStatusMapping[name]}
-											<button on:click={() => cancelUpload(name)} class="delete"></button>
-										{/if}
-									</td>
+									{#if !uploadStatusMapping[name]}
+										<td>
+											<div style="width: fit-content">
+												<button on:click={() => cancelUpload(name)} class="button is-small is-link"
+													>Cancel Upload</button
+												>
+											</div>
+										</td>
+									{/if}
 								{/if}
 							</tr>
 						{/each}
@@ -607,9 +621,9 @@
 		</div>
 	{/if}
 
-	<div class="columns mt-5 is-align-items-center">
+	<div class="columns is-mobile mt-5">
 		<form
-			class="column is-fullwidth is-flex is-justify-content-left"
+			class="column is-flex is-justify-content-start"
 			method="POST"
 			on:submit={() => {
 				isUploading = true;
@@ -625,24 +639,26 @@
 			}}
 		>
 			<input class="input" type="hidden" name="SelectedFiles" bind:value={selectedFilesList} />
-			<div class="pl-0 control column is-one-fifth">
-				<button
-					class="button is-medium is-primary {isUploading ? 'is-loading' : ''}"
-					disabled={uploadDisabled || isUploading}
-					type="submit"
-				>
-					<span class="icon">
-						<i class="fa-solid fa-cloud-arrow-up" />
-					</span>
-					<span>Upload</span>
-				</button>
-			</div>
+			<button
+				class="button is-medium is-primary {isUploading ? 'is-loading' : ''}"
+				disabled={uploadDisabled || isUploading}
+				type="submit"
+			>
+				<span class="icon">
+					<i class="fa-solid fa-cloud-arrow-up" />
+				</span>
+				<span>Upload</span>
+			</button>
 		</form>
-		<button
-			on:click={removeAllFiles}
-			disabled={filesToUpload.length < 1 || !userIsSignedIn || isUploading}
-			class="button is-medium is-link">Clear all</button
-		>
+		<div class="column is-flex is-justify-content-end">
+			<button
+				on:click={removeAllFiles}
+				disabled={filesToUpload.length < 1 || !userIsSignedIn || isUploading}
+				class="button is-medium is-link is-fullwidth-mobile"
+			>
+				Clear all
+			</button>
+		</div>
 	</div>
 
 	{#if showErrorMessages}